+++ app/build.gradle
@@ -1,49 +1,86 @@
-plugins {
-    id 'com.android.application'
+/*
+ *  Copyright 2018 Exotel Techcom Pvt Ltd
+ */
+
+buildscript {
+    repositories {
+        mavenCentral()
+        maven { url 'https://maven.fabric.io/public' }
+        maven { url 'https://jitpack.io' }
+    }
+
+    dependencies {
+        // These docs use an open ended version so that our plugin
+        // can be updated quickly in response to Android tooling updates
+
+        // We recommend changing it to the latest version from our changelog:
+        // https://docs.fabric.io/android/changelog.html#fabric-gradle-plugin
+        classpath 'io.fabric.tools:gradle:1.+'
+    }
+}
+
+apply plugin: 'com.android.application'
+
+apply plugin: 'io.fabric'
+
+configurations.all {
+    // Check for updates every build
+    resolutionStrategy.cacheChangingModulesFor 0, 'seconds'
+}
+
+
+repositories {
+    mavenCentral()
+    maven { url 'https://maven.fabric.io/public' }
+//    maven { url 'https://jitpack.io' }
+    maven { url 'https://mymavenrepo.com/repo/e01iL1W9Gdxw73PGI9LA/' }
 }
 
 android {
-    namespace 'com.example.orchestrationSampleApp'
-    compileSdk 32
-
+    compileSdkVersion 28
+    buildToolsVersion '27.0.3'
     defaultConfig {
-        applicationId "com.example.orchestrationSampleApp"
-        minSdk 22
-        targetSdk 32
+        applicationId "com.exotel.verificationdemoapp"
+        minSdkVersion 16
+        targetSdkVersion 26
         versionCode 1
-        versionName "1.0"
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        versionName "1.6.0"
     }
-
+    lintOptions {
+        abortOnError false
+    }
     buildTypes {
+        debug {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+            debuggable true
+        }
         release {
             minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+            debuggable false
         }
-    }
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-    buildFeatures {
-        viewBinding true
     }
 }
 
 dependencies {
-
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    implementation 'com.squareup.okhttp3:okhttp:3.4.1'
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+    androidTestImplementation('com.android.support.test.espresso:espresso-core:2.2.2', {
+        exclude group: 'com.android.support', module: 'support-annotations'
+    })
+    implementation 'com.android.support:appcompat-v7:25.3.0'
+    implementation 'org.bitbucket.Exotel:exoverify:2.0.0'
+    implementation 'com.squareup.okhttp3:okhttp:3.6.0'
     implementation 'com.google.code.gson:gson:2.8.0'
     implementation 'com.googlecode.libphonenumber:libphonenumber:8.8.3'
     implementation 'dnsjava:dnsjava:2.1.6'
-    testImplementation 'junit:junit:4.12'
-    implementation 'org.bitbucket.Exotel:exoverify:1.7.0'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+//    implementation project(':exotelVerificationLibrary')
+//    implementation('com.exotel:verification:0.1.8@aar') {
+//        transitive = true;
+//        changing = true;
+//    }
+    implementation('com.crashlytics.sdk.android:crashlytics:2.6.8@aar') {
+        transitive = true;
+    }
     implementation 'pl.droidsonroids.gif:android-gif-drawable:1.2.7'
-
 }
