+++ build.gradle
@@ -1,5 +1,37 @@
+/*
+ * Copyright 2017 Exotel India Pvt Ltd
+ *
+ */
+
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
-plugins {
-    id 'com.android.application' version '7.3.1' apply false
-    id 'com.android.library' version '7.3.1' apply false
+
+buildscript {
+    repositories {
+        jcenter()
+        maven {
+            url "https://maven.google.com/"
+        }
+        google()
+    }
+    dependencies {
+        classpath 'com.android.tools.build:gradle:3.3.2'
+
+        // NOTE: Do not place your application dependencies here; they belong
+        // in the individual module build.gradle files
+    }
+}
+
+allprojects {
+    repositories {
+        google()
+        jcenter()
+        maven {
+            url 'https://jitpack.io'
+            credentials  { username 'jp_etcct006nc8pkd0ntra0n5uk9k' }
+        }
+    }
+}
+
+task clean(type: Delete) {
+    delete rootProject.buildDir
 }
